minor.
git-svn-id: http://svn.cmsmadesimple.org/svn/cmsmadesimple/trunk@10746 3d254a34-79dc-0310-9e5f-be208747d8a0

+++ modules/ModuleManager/action.setprefs.php
@@ -44,7 +44,7 @@ if( isset($config['developer_mode']) && isset($params['reseturl']) ) {
   $this->SetMessage($this->Lang('msg_urlreset'));
   $this->RedirectToAdminTab();
 }
-if( isset($params['dl_chunksize']) ) $this->SetPreference('dl_chunksize',trim($params['dl_chunksize']));
+if( isset($params['dl_chunksize']) ) $this->SetPreference('dl_chunksize',(int)trim($params['dl_chunksize']));
 $latestdepends = (int)get_parameter_value($params,'latestdepends');
 $this->SetPreference('latestdepends',$latestdepends);
